requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
unstructured
unstructured[local-inference]
python-dotenv
supabase
pgvector
openai
arq
redis
httpx
numpy


